app: move view 'solve_captcha' to 'controllers'
