+++ decaptcha/controllers/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import absolute_import
+
+from .solve import solve_captcha
